UIQM-132 Move getQuickMarcRecordStatus function into a separate file
