+++ src/QuickMarcEditor/QuickMarcDuplicateWrapper.js
@@ -14,10 +14,9 @@ import {
 } from '@folio/stripes/components';
 
 import QuickMarcEditor from './QuickMarcEditor';
+import getQuickMarcRecordStatus from './getQuickMarcRecordStatus';
 import {
   QUICK_MARC_ACTIONS,
-  QM_RECORD_STATUS_TIMEOUT,
-  QM_RECORD_STATUS_BAIL_TIME,
 } from './constants';
 import { MARC_TYPES } from '../common/constants';
 import {
@@ -64,54 +63,6 @@ const QuickMarcDuplicateWrapper = ({
     onClose();
   };
 
-  const getQuickMarcRecordStatus = (qmRecordId) => {
-    const maxRequestAttempts = QM_RECORD_STATUS_BAIL_TIME / QM_RECORD_STATUS_TIMEOUT;
-    let requestCount = 1;
-    let intervalId;
-
-    function makeRequest() {
-      mutator.quickMarcRecordStatus.GET({ params: { qmRecordId } })
-        .then(({ externalId, status }) => {
-          if (status === 'ERROR') {
-            clearInterval(intervalId);
-            showCallout({
-              messageId: 'ui-quick-marc.record.saveNew.error',
-              type: 'error',
-            });
-          }
-
-          if (status === 'IN_PROGRESS') {
-            if (requestCount === maxRequestAttempts) {
-              clearInterval(intervalId);
-              showCallout({ messageId: 'ui-quick-marc.record.saveNew.delay' });
-            } else {
-              requestCount++;
-            }
-          }
-
-          if (externalId !== null && status === 'CREATED') {
-            clearInterval(intervalId);
-            showCallout({ messageId: 'ui-quick-marc.record.saveNew.success' });
-
-            history.push({
-              pathname: `/inventory/view/${externalId}`,
-              search: location.search,
-            });
-          }
-        })
-        .catch(() => {
-          showCallout({
-            messageId: 'ui-quick-marc.record.saveNew.error',
-            type: 'error',
-          });
-        });
-    }
-
-    makeRequest();
-
-    intervalId = setInterval(makeRequest, QM_RECORD_STATUS_TIMEOUT);
-  };
-
   const onSubmit = useCallback(async (formValues) => {
     const autopopulatedFormValues = autopopulateSubfieldSection(removeFieldsForDuplicate(formValues), marcType);
     const formValuesForDuplicate = cleanBytesFields(autopopulatedFormValues, initialValues, marcType);
@@ -132,7 +83,13 @@ const QuickMarcDuplicateWrapper = ({
           search: location.search,
         });
 
-        getQuickMarcRecordStatus(qmRecordId);
+        getQuickMarcRecordStatus({
+          mutator,
+          qmRecordId,
+          showCallout,
+          history,
+          location,
+        });
       })
       .catch(async (errorResponse) => {
         let messageId;
